Embeddings optimization and other changes
Embeddings can be stored and loaded providing a specific
optimization strategy (generated and storage embeddings).

Logging initialization has been refactored in a way which the
configuration is propagated instead of initialized for each file.

Storage embeddings, which are expected to be monotonic with the
sentences, now can be defined if they are unique or there is an
embedding per sentence even if had been already seen previously.